Mark up IDL from constrainable patterns as non-normative
The Capabitilies, Settings, Constraints and ConstraintSet dictionataries are not meant to be directly implemented - they document the Constrainable pattern.

To avoid having them be considered as part of the normative surface of the spec, they're marked-up with the example class - similar to what we do with ConstrainablePattern.

This in particular helps with automatically extracted IDL fragments as done by webref / reffy https://github.com/w3c/webref/